requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
alpaca-trade-api==2.3.0
aniso8601==9.0.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astor==0.8.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
binance-futures-connector==3.2.0
bitcoinaddress==0.1.5
blankly==1.18.19b0
bleach==5.0.1
blessed==1.19.1
bokeh==2.4.3
boto3==1.24.91
botocore==1.27.91
bt==0.2.9
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==8.6.0
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
colorful==0.5.4
constantly==15.1.0
contourpy==1.0.5
cp==2020.12.3
cryptography==38.0.3
crytic-compile==0.2.4
cycler==0.11.0
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.10.0
dateparser==1.1.3
db-dtypes==1.0.4
debugpy==1.6.3
decorator==5.1.1
DeFiLlama==1.1.0
DeFiLlama-Curl==1.2
defillama2==0.5
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
distributed==2022.10.0
Django==4.1.2
ecdsa==0.17.0
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
ffn==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.37.4
fredapi==0.5.0
frozenlist==1.3.1
fsspec==2022.11.0
future==0.18.2
gast==0.5.3
gcsfs==2022.11.0
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gpustat==1.0.0
grpcio==1.50.0
grpcio-status==1.50.0
h5py==3.7.0
HeapDict==1.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jaraco.functools==3.5.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
mock-open==1.4.0
modin==0.16.1
modin-spreadsheet==0.1.2
more-itertools==9.0.0
msgpack==1.0.3
multidict==6.0.2
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
newnewtulipy==0.4.6.4
notebook==6.4.12
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.0
opencensus-context==0.1.3
packaging==21.3
pandas==1.5.0
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
plotly-express==0.4.1
plumbum==1.8.0
prettytable==3.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.8
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycoingecko==3.0.0
pycparser==2.21
pycurl==7.45.1
pydantic==1.9.1
pydata-google-auth==1.4.0
Pygments==2.13.0
pyOpenSSL==22.1.0
pyorbital==1.7.3
pyparsing==3.0.9
PyPrind==2.11.3
pyrsistent==0.18.1
pysha3==1.0.2
PySimpleGUI==4.60.4
-e git+https://github.com/Homebrew/brew@759ab2d0bdc6b0d1f810d6e9351c834d027809f7#egg=python_binance
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
questionary==1.10.0
ray==2.0.0
redis==3.5.3
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
rpyc==4.1.5
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
service-identity==21.1.0
shroomdk==1.0.1
six==1.16.0
slither-analyzer==0.9.0
smart-open==6.2.0
solc-select==1.0.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
statsmodels==0.13.2
tabulate==0.9.0
tblib==1.7.0
tenacity==8.1.0
tensorboard==1.12.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl
termcolor==2.0.1
termcolor-whl==1.1.2
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
tornado==6.1
traitlets==5.4.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ubtsl==0.0.1
ujson==5.5.0
unicorn-binance-local-depth-cache==0.7.3
unicorn-binance-rest-api==1.5.0
unicorn-binance-suite==1.1.0
unicorn-binance-trailing-stop-loss==0.8.0
unicorn-binance-websocket-api==1.41.0
unicorn-fy==0.12.2
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarl==1.8.1
yaspin==2.2.0
zict==2.2.0
zope.interface==5.5.2
